Added unit tests and integration tests

+++ src/main/java/com/divjazz/recommendic/article/dto/ArticleSearchDTO.java
@@ -1,10 +1,5 @@
 package com.divjazz.recommendic.article.dto;
 
-import com.divjazz.recommendic.user.model.userAttributes.UserName;
-
-import java.time.LocalDateTime;
-import java.util.Set;
-
 public record ArticleSearchDTO(
         long id,
         String title,
@@ -13,11 +8,11 @@ public record ArticleSearchDTO(
         String authorLastName,
         String publishedAt,
         String[] tags,
-        float rank,
+        double rank,
         String highlighted,
         long upvotes,
-        int numberOfComment,
-        int reads,
+        long numberOfComment,
+        long reads,
         long total
 ) {
 }
